Second commit - refactoring taskrun_test.go into five additional files - one for each host provision type and another file for all the various mocks and the helper methods for the tests. Some additional tests were added where complete fail scenarios were missing (in static host pools and in dynamic pools).
